# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/ensayos": {
      "post": {
        "operationId": "EnsayosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnsayoDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "nombreEnsayo": "Ensayo de Trigo 2025",
                    "versionProtocolo": "v1",
                    "responsable": "Ing. Agr. Juan Pérez",
                    "provincia": "Buenos Aires",
                    "departamento": "Tres Arroyos",
                    "establecimiento": "La Estancia",
                    "lote": "Lote 5",
                    "latitud": -38.373737,
                    "longitud": -60.273737,
                    "cultivoEspecie": "Trigo",
                    "cultivoVariedad": "Baguette 620",
                    "tipoSiembra": "A chorrillo",
                    "distSurcosCm": 17.5,
                    "fechaSiembra": "2025-06-15"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ensayo creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ensayo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un ensayo",
        "tags": [
          "ensayos"
        ]
      },
      "get": {
        "operationId": "EnsayosController_findAll",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Búsqueda por nombre/version",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campo para ordenar (ej: id, nombreEnsayo)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [],
                    "meta": {
                      "total": 0,
                      "page": 1,
                      "limit": 10,
                      "pageCount": 0
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos los ensayos (paginado y filtrado)",
        "tags": [
          "ensayos"
        ]
      }
    },
    "/api/v1/ensayos/{id}": {
      "get": {
        "operationId": "EnsayosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ensayo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener un ensayo por ID",
        "tags": [
          "ensayos"
        ]
      },
      "patch": {
        "operationId": "EnsayosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEnsayoDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "responsable": "María López",
                    "provincia": "Córdoba",
                    "fechaSiembra": "2025-06-20"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ensayo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar un ensayo por ID",
        "tags": [
          "ensayos"
        ]
      },
      "delete": {
        "operationId": "EnsayosController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deleted": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un ensayo por ID",
        "tags": [
          "ensayos"
        ]
      }
    },
...
    "/api/v1/datos-cosecha": {
      "post": {
        "operationId": "DatosCosechaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatosCosechaDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "parcelaId": 1,
                    "fechaCosecha": "2025-05-30",
                    "humedadPct": 14.2,
                    "kgHaCorregido": 5200.5,
                    "gie": 123.45,
                    "observaciones": "Trilla sin inconvenientes"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Datos de cosecha creados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatosCosecha"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear registro de datos de cosecha",
        "tags": [
          "datos-cosecha"
        ]
      },
      "get": {
        "operationId": "DatosCosechaController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatosCosecha"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar datos de cosecha",
        "tags": [
          "datos-cosecha"
        ]
      }
    },
    "/api/v1/datos-cosecha/{id}": {
      "get": {
        "operationId": "DatosCosechaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatosCosecha"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener datos de cosecha por ID",
        "tags": [
          "datos-cosecha"
        ]
      },
      "patch": {
        "operationId": "DatosCosechaController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDatosCosechaDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "humedadPct": 13.8,
                    "observaciones": "Secado adicional realizado"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatosCosecha"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar datos de cosecha",
        "tags": [
          "datos-cosecha"
        ]
      },
      "delete": {
        "operationId": "DatosCosechaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deleted": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar datos de cosecha",
        "tags": [
          "datos-cosecha"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Ensayos Agronómicos",
    "description": "Documentación de la API para el sistema de gestión de ensayos",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateEnsayoDto": {
        "type": "object",
        "properties": {
          "nombreEnsayo": {
            "type": "string",
            "description": "Nombre del ensayo",
            "example": "Ensayo de Trigo 2025",
            "maxLength": 255
          },
          "versionProtocolo": {
            "type": "object",
            "description": "Versión del protocolo",
            "example": "v1",
            "maxLength": 20
          },
          "responsable": {
            "type": "object",
            "description": "Responsable del ensayo",
            "example": "Ing. Agr. Juan Pérez",
            "maxLength": 100
          },
          "provincia": {
            "type": "object",
            "description": "Provincia",
            "example": "Buenos Aires",
            "maxLength": 100
          },
          "departamento": {
            "type": "object",
            "description": "Departamento",
            "example": "Tres Arroyos",
            "maxLength": 100
          },
          "establecimiento": {
            "type": "object",
            "description": "Establecimiento",
            "example": "La Estancia",
            "maxLength": 100
          },
          "lote": {
            "type": "object",
            "description": "Lote",
            "example": "Lote 5",
            "maxLength": 50
          },
          "latitud": {
            "type": "object",
            "description": "Latitud en grados decimales",
            "example": -38.373737
          },
          "longitud": {
            "type": "object",
            "description": "Longitud en grados decimales",
            "example": -60.273737
          },
          "cultivoEspecie": {
            "type": "object",
            "description": "Especie del cultivo",
            "example": "Trigo",
            "maxLength": 100
          },
          "cultivoVariedad": {
            "type": "object",
            "description": "Variedad del cultivo",
            "example": "Baguette 620",
            "maxLength": 100
          },
          "tipoSiembra": {
            "type": "object",
            "description": "Tipo de siembra",
            "example": "A chorrillo",
            "maxLength": 50
          },
          "distSurcosCm": {
            "type": "object",
            "description": "Distancia entre surcos (cm)",
            "example": 17.5
          },
          "fechaSiembra": {
            "type": "object",
            "description": "Fecha de siembra (YYYY-MM-DD)",
            "example": "2025-06-15"
          }
        },
        "required": [
          "nombreEnsayo"
        ]
      },
      "Ensayo": {
        "type": "object",
        "properties": {}
      },
      "UpdateEnsayoDto": {
        "type": "object",
        "properties": {
          "nombreEnsayo": {
            "type": "string",
            "description": "Nombre del ensayo",
            "example": "Ensayo de Trigo 2025",
            "maxLength": 255
          },
          "versionProtocolo": {
            "type": "object",
            "description": "Versión del protocolo",
            "example": "v1",
            "maxLength": 20
          },
          "responsable": {
            "type": "object",
            "description": "Responsable del ensayo",
            "example": "Ing. Agr. Juan Pérez",
            "maxLength": 100
          },
          "provincia": {
            "type": "object",
            "description": "Provincia",
            "example": "Buenos Aires",
            "maxLength": 100
          },
          "departamento": {
            "type": "object",
            "description": "Departamento",
            "example": "Tres Arroyos",
            "maxLength": 100
          },
          "establecimiento": {
            "type": "object",
            "description": "Establecimiento",
            "example": "La Estancia",
            "maxLength": 100
          },
          "lote": {
            "type": "object",
            "description": "Lote",
            "example": "Lote 5",
            "maxLength": 50
          },
          "latitud": {
            "type": "object",
            "description": "Latitud en grados decimales",
            "example": -38.373737
          },
          "longitud": {
            "type": "object",
            "description": "Longitud en grados decimales",
            "example": -60.273737
          },
          "cultivoEspecie": {
            "type": "object",
            "description": "Especie del cultivo",
            "example": "Trigo",
            "maxLength": 100
          },
          "cultivoVariedad": {
            "type": "object",
            "description": "Variedad del cultivo",
            "example": "Baguette 620",
            "maxLength": 100
          },
          "tipoSiembra": {
            "type": "object",
            "description": "Tipo de siembra",
            "example": "A chorrillo",
            "maxLength": 50
          },
          "distSurcosCm": {
            "type": "object",
            "description": "Distancia entre surcos (cm)",
            "example": 17.5
          },
          "fechaSiembra": {
            "type": "object",
            "description": "Fecha de siembra (YYYY-MM-DD)",
            "example": "2025-06-15"
          }
        }
      },
...
      "CreateDatosCosechaDto": {
        "type": "object",
        "properties": {
          "parcelaId": {
            "type": "number",
            "example": 1,
            "description": "ID de la parcela"
          },
          "fechaCosecha": {
            "type": "object",
            "example": "2025-05-30"
          },
          "humedadPct": {
            "type": "object",
            "example": 14.2,
            "description": "Humedad (%)"
          },
          "kgHaCorregido": {
            "type": "object",
            "example": 5200.5,
            "description": "kg/ha corregido"
          },
          "gie": {
            "type": "object",
            "example": 123.45
          },
          "observaciones": {
            "type": "object",
            "example": "Observaciones de la trilla",
            "maxLength": 65535
          }
        },
        "required": [
          "parcelaId"
        ]
      },
      "DatosCosecha": {
        "type": "object",
        "properties": {}
      },
      "UpdateDatosCosechaDto": {
        "type": "object",
        "properties": {
          "parcelaId": {
            "type": "number",
            "example": 1,
            "description": "ID de la parcela"
          },
          "fechaCosecha": {
            "type": "object",
            "example": "2025-05-30"
          },
          "humedadPct": {
            "type": "object",
            "example": 14.2,
            "description": "Humedad (%)"
          },
          "kgHaCorregido": {
            "type": "object",
            "example": 5200.5,
            "description": "kg/ha corregido"
          },
          "gie": {
            "type": "object",
            "example": 123.45
          },
          "observaciones": {
            "type": "object",
            "example": "Observaciones de la trilla",
            "maxLength": 65535
          }
        }
      },
...
    }
  }
}